+++ nitrogen/test/src/file_system_test.dart
@@ -1,4 +1,4 @@
-import 'package:nitrogen_types/nitrogen_types.dart';
+import 'package:nitrogen_types/assets.dart';
 import 'package:test/test.dart';
 
 import 'package:nitrogen/src/file_system.dart';