+++ nitrogen/test/src/generators/asset_generator_test.dart
@@ -1,13 +1,107 @@
 import 'package:code_builder/code_builder.dart';
 import 'package:dart_style/dart_style.dart';
-import 'package:nitrogen_types/nitrogen_types.dart';
+import 'package:nitrogen_types/assets.dart';
 import 'package:test/test.dart';
 
 import 'package:nitrogen/src/file_system.dart';
 import 'package:nitrogen/src/generators/asset_generator.dart';
 
-const _classes = r'''
-import 'package:nitrogen_types/nitrogen_types.dart';
+const _classesDocs =  r'''
+import 'package:nitrogen_types/assets.dart';
+
+// GENERATED CODE - DO NOT MODIFY BY HAND
+//
+// **************************************************************************
+// nitrogen
+// **************************************************************************
+//
+// ignore_for_file: type=lint
+// ignore_for_file: deprecated_member_use
+
+/// Contains the assets and nested directories in the `path/to/directory` directory.
+///
+/// Besides the assets and nested directories, it also provides a [contents] for querying the assets in the current
+/// directory.
+///
+/// To convert an asset into a widget, [call the asset like a function](https://dart.dev/language/callable-objects].
+/// ```dart
+/// final widget = $PrefixPathToDirectory.path.to.asset();
+/// ```
+///
+/// The `call(...)` functions are provided by extensions. By default, only the [ImageAsset] extension is bundled.
+///
+/// 3rd party packages are supported via 'extension' packages. `extension` packages contain an `extension` that provide a
+/// `call(...)` function that transforms an `Asset` into a 3rd party type.
+///
+/// | Type              | Package       | Extension Package      | Version                                                                                                        |
+/// |-------------------|---------------|------------------------|----------------------------------------------------------------------------------------------------------------|
+/// | SVG images        | `flutter_svg` | `nitrogen_flutter_svg` | [![Pub Dev](https://img.shields.io/pub/v/nitrogen_flutter_svg)](https://pub.dev/packages/nitrogen_flutter_svg) |
+/// | Lottie animations | `lottie`      | `nitrogen_lottie`      | [![Pub Dev](https://img.shields.io/pub/v/nitrogen_lottie)](https://pub.dev/packages/nitrogen_lottie)           |
+class $PrefixPathToDirectory {
+  const $PrefixPathToDirectory();
+
+  /// The `path/to/directory/subdirectory` directory.
+  static $PrefixPathToDirectorySubdirectory get subdirectory => const $PrefixPathToDirectorySubdirectory();
+
+  /// The `path/to/directory/bar.txt`.
+  static GenericAsset get bar => const GenericAsset(
+        'test_package',
+        'bar',
+        'path/to/directory/bar.txt',
+      );
+
+  /// The contents of this directory.
+  static Map<String, Asset> get contents => const {
+        'bar': const GenericAsset(
+          'test_package',
+          'bar',
+          'path/to/directory/bar.txt',
+        ),
+      };
+}
+
+/// Contains the assets and nested directories in the `path/to/directory/subdirectory` directory.
+///
+/// Besides the assets and nested directories, it also provides a [contents] for querying the assets in the current
+/// directory.
+///
+/// To convert an asset into a widget, [call the asset like a function](https://dart.dev/language/callable-objects].
+/// ```dart
+/// final widget = $PrefixPathToDirectorySubdirectory.path.to.asset();
+/// ```
+///
+/// The `call(...)` functions are provided by extensions. By default, only the [ImageAsset] extension is bundled.
+///
+/// 3rd party packages are supported via 'extension' packages. `extension` packages contain an `extension` that provide a
+/// `call(...)` function that transforms an `Asset` into a 3rd party type.
+///
+/// | Type              | Package       | Extension Package      | Version                                                                                                        |
+/// |-------------------|---------------|------------------------|----------------------------------------------------------------------------------------------------------------|
+/// | SVG images        | `flutter_svg` | `nitrogen_flutter_svg` | [![Pub Dev](https://img.shields.io/pub/v/nitrogen_flutter_svg)](https://pub.dev/packages/nitrogen_flutter_svg) |
+/// | Lottie animations | `lottie`      | `nitrogen_lottie`      | [![Pub Dev](https://img.shields.io/pub/v/nitrogen_lottie)](https://pub.dev/packages/nitrogen_lottie)           |
+class $PrefixPathToDirectorySubdirectory {
+  const $PrefixPathToDirectorySubdirectory();
+
+  /// The `path/to/directory/subdirectory/foo.png`.
+  ImageAsset get foo => const ImageAsset(
+        'test_package',
+        'foo',
+        'path/to/directory/subdirectory/foo.png',
+      );
+
+  /// The contents of this directory.
+  Map<String, Asset> get contents => const {
+        'foo': const ImageAsset(
+          'test_package',
+          'foo',
+          'path/to/directory/subdirectory/foo.png',
+        ),
+      };
+}
+''';
+
+const _classesNoDocs = r'''
+import 'package:nitrogen_types/assets.dart';
 
 // GENERATED CODE - DO NOT MODIFY BY HAND
 //
@@ -178,29 +272,36 @@ void main() {
     const GenericAsset('test_package', 'bar', 'path/to/directory/bar.txt'),
   );
 
-  test('AssetGenerator', () {
-    final generator = AssetGenerator('Prefix', directory, {});
-    expect(formatter.format(generator.generate()), _classes);
+  group('AssetGenerator', () {
+    test('docs', () {
+      final generator = AssetGenerator('Prefix', directory, {}, docs: true);
+      expect(formatter.format(generator.generate()), _classesDocs);
+    });
+
+    test('no docs', () {
+      final generator = AssetGenerator('Prefix', directory, {}, docs: false);
+      expect(formatter.format(generator.generate()), _classesNoDocs);
+    });
   });
 
   group('AssetClass', () {
     group('generate(...)', () {
       test('non static', () {
-        const basic = AssetClass(directories: AssetDirectoryExpressions('Prefix'));
+        const basic = AssetClass(docs: false, directories: AssetDirectoryExpressions('Prefix'));
         final type = basic.generate(directory).build();
 
         expect(formatter.format(type.accept(emitter).toString()), formatter.format(_nonStaticAssetClass));
       });
 
       test('static', () {
-        const basic = AssetClass(directories: AssetDirectoryExpressions('Prefix'));
+        const basic = AssetClass(docs: false, directories: AssetDirectoryExpressions('Prefix'));
         final type = basic.generate(directory, static: true).build();
 
         expect(formatter.format(type.accept(emitter).toString()), formatter.format(_staticAssetClass));
       });
 
       test('excluded', () {
-        final basic = AssetClass(directories: const AssetDirectoryExpressions('Prefix'), excluded: { subdirectory });
+        final basic = AssetClass(docs: false, directories: const AssetDirectoryExpressions('Prefix'), excluded: { subdirectory });
         final type = basic.generate(directory).build();
 
         expect(formatter.format(type.accept(emitter).toString()), formatter.format(_excludedAssetClass));
@@ -211,21 +312,21 @@ void main() {
   group('BasicAssetClass', () {
     group('generate(...)', () {
       test('non static', () {
-        const basic = BasicAssetClass(directories: AssetDirectoryExpressions('Prefix'));
+        const basic = BasicAssetClass(docs: false, directories: AssetDirectoryExpressions('Prefix'));
         final type = basic.generate(directory).build();
 
         expect(formatter.format(type.accept(emitter).toString()), formatter.format(_nonStaticBasicClass));
       });
 
       test('static', () {
-        const basic = BasicAssetClass(directories: AssetDirectoryExpressions('Prefix'));
+        const basic = BasicAssetClass(docs: false, directories: AssetDirectoryExpressions('Prefix'));
         final type = basic.generate(directory, static: true).build();
 
         expect(formatter.format(type.accept(emitter).toString()), formatter.format(_staticBasicClass));
       });
 
       test('excluded', () {
-        final basic = BasicAssetClass(directories: const AssetDirectoryExpressions('Prefix'), excluded: { subdirectory });
+        final basic = BasicAssetClass(docs: false, directories: const AssetDirectoryExpressions('Prefix'), excluded: { subdirectory });
         final type = basic.generate(directory).build();
 
         expect(formatter.format(type.accept(emitter).toString()), formatter.format(_excludedBasicClass));