+++ nitrogen/test/src/walker_test.dart
@@ -1,5 +1,5 @@
 import 'package:build/build.dart';
-import 'package:nitrogen_types/nitrogen_types.dart';
+import 'package:nitrogen_types/assets.dart';
 import 'package:test/test.dart';
 
 import 'package:nitrogen/src/file_system.dart';