+++ nitrogen_types/lib/assets.dart
@@ -1,0 +1,4 @@
+/// This library exports only the assets. It is used by Nitrogen since builders cannot import Flutter.
+library nitrogen_types.assets;
+
+export 'src/assets.dart';